refactor: split matches.ts into schedule/results modules
